Portfolio Section is Compplete

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -7,6 +7,7 @@ use App\Models\Category;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\DataTables\CategoryDataTable;
+use App\Models\PortfolioItem;
 
 class CategoryController extends Controller
 {
@@ -111,6 +112,14 @@ class CategoryController extends Controller
     public function destroy($id)
     {
          $category = Category::findOrFail($id);
-         $category->delete();
+
+         $hasItem = PortfolioItem::where('category_id', $category->id)->count();
+         if($hasItem == 0){
+            $category->delete();
+            return true;
+         }
+
+         return response(['status'=> 'error']);
+        
     }
 }
